Side-by-side comparison of results

+++ src/session.rs
@@ -9,16 +9,16 @@ pub fn cluster(conf: &Config) -> Cluster {
         cluster.set_contact_points(addr).unwrap();
     }
     cluster
-        .set_core_connections_per_host(conf.connections)
+        .set_core_connections_per_host(conf.connections as u32)
         .unwrap();
     cluster
-        .set_max_connections_per_host(conf.connections)
+        .set_max_connections_per_host(conf.connections as u32)
         .unwrap();
     cluster
         .set_queue_size_event(conf.parallelism as u32)
         .unwrap();
     cluster.set_queue_size_io(conf.parallelism as u32).unwrap();
-    cluster.set_num_threads_io(conf.threads).unwrap();
+    cluster.set_num_threads_io(conf.threads as u32).unwrap();
     cluster.set_connect_timeout(time::Duration::seconds(5));
     cluster.set_load_balance_round_robin();
     cluster
